requirements/base.txt
django==3.0.8
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.8.5
factory-boy==2.12.0
plotly==4.9.0
numpy==1.19.1
